+++ Source/ilyvion.LoadingProgress/DeepProfiler_Start_Patches.cs
@@ -1,3 +1,5 @@
+using System.Diagnostics;
+
 namespace ilyvion.LoadingProgress;
 
 [HarmonyPatch(typeof(DeepProfiler), nameof(DeepProfiler.Start))]
@@ -5,6 +7,15 @@ internal static class DeepProfiler_Start_Patches
 {
     private static void Prefix(string label)
     {
-        LoadingProgressWindow.CurrentLoadingActivity = label;
+        if (label == null)
+        {
+            MethodBase method = new StackTrace().GetFrame(2).GetMethod();
+            var mod = Utilities.FindModByAssembly(method.DeclaringType.Assembly);
+            LoadingProgressMod.Warning($"Why is {method.DeclaringType.FullName}.{method.Name} from {mod?.Name ?? "{unknown}"} calling DeepProfiler.Start (and by extension our patch) with null?! Stop it.");
+        }
+        else
+        {
+            LoadingProgressWindow.CurrentLoadingActivity = label;
+        }
     }
 }