+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+- Mods have apparently decided to call DeepProfiler.Start with null. We didn't expect this. Now we're handling it.
+
 ## [0.9.0] - 2025-08-22
 
 ### Added